make the tests compile 	* mln/convert/to_image.hh: replace box2d by box_<P> to handle 1d 	* mln/debug/println.hh: stupid error 	* tests/convert_to_vec_p.cc: stupid error 	* tests/w_window1d_int.cc: rectangle1d ??? replaced by segment1d
git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@1091 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/tests/convert_to_vec_p.cc
@@ -48,7 +48,7 @@ int main()
   win::segment1d win1d(5);
   vec1d v1(convert::to_vec_p(win1d, p1));
 
-  for (int i=0; i < v1.npoints(); i++)
+  for (size_t i=0; i < v1.npoints(); i++)
     std::cout << (v1[i]);
 
   std::cout << "\n";
@@ -58,7 +58,7 @@ int main()
   win::rectangle2d win2d(3, 3);
   vec2d v2(convert::to_vec_p(win2d, p2));
 
-  for (int i=0; i < v2.npoints(); i++)
+  for (size_t i=0; i < v2.npoints(); i++)
     std::cout << (v2[i]);
 
 }
